requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.32.0
google-generativeai==0.3.2
ollama==0.1.7